requirements.txt
Django==5.1.1
django-environ==0.11.2
psycopg[binary]==3.2.12
channels==4.1.0
channels-redis==4.2.0
redis==5.0.8
daphne==4.1.2
requests==2.32.3
Pillow==10.4.0
